Compile the strings markup at import time
- tools: Add _markupparser module, and use it from get-strings-file
  script to compile the markup at import time.

- eosclubhouse clubhouse & utils: Remove SimpleMarkupParser class and
  its calls.

- tests: Remove test for SimpleMarkupParser. They were migrated to
  doctests in the _markupparser tools module.

- data: Import the strings so they get the markup compiled.

https://phabricator.endlessm.com/T29020
