+++ data/quests_strings/artsnapshot.csv
@@ -1,5 +1,5 @@
 ARTSNAPSHOT_QUEST_NAME,Let's Be (photo) Realistic,,,,
-ARTSNAPSHOT_WELCOME,"I'm sure you've heard the phrase *""A picture is worth a thousand words.""*
+ARTSNAPSHOT_WELCOME,"I'm sure you've heard the phrase <b>""A picture is worth a thousand words.""</b>
 
 But have you ever tried to figure out what that means, or how you'd create that effect?",estelle,,,
 ARTSNAPSHOT_QUEST_SUBTITLE,What you'll do,,,,
