+++ data/quests_strings/blendmonster2.csv
@@ -2,12 +2,12 @@ BLENDMONSTER2_QUEST_NAME,Monster Smoothie - Assembling the Ingredients,,,,
 BLENDMONSTER2_QUEST_SUBTITLE,What you'll do,,,,
 BLENDMONSTER2_QUEST_DESCRIPTION,"In part 2 of this quest, it's time to dive into modeling one of the characters in your scene.",,,,
 BLENDMONSTER2_QUIT,See you later!,,,,
-BLENDMONSTER2_NOTINSTALLED,"Uh-oh! Looks like `Blender` isn't installed!
+BLENDMONSTER2_NOTINSTALLED,"Uh-oh! Looks like <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Blender</span></tt> isn't installed!
 
 You need to download it from the App Center if you'd like to learn how to use it.
 
-Click the icon named `App Center` on the Desktop, then search for `Blender`, and install it. ",daemon,,,
-BLENDMONSTER2_INFO_1,"Welcome back, *{{user_name}}*! I'd recommend doing the previous parts of this quest if you haven't already. If you have, or you're forging ahead, plug in your mouse and let's get modeling!",estelle,,,
+Click the icon named <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">App Center</span></tt> on the Desktop, then search for <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Blender</span></tt>, and install it. ",daemon,,,
+BLENDMONSTER2_INFO_1,"Welcome back, <b>{{user_name}}</b>! I'd recommend doing the previous parts of this quest if you haven't already. If you have, or you're forging ahead, plug in your mouse and let's get modeling!",estelle,,,
 BLENDMONSTER2_INFO_2,The easiest way to follow along with these tutorials is to adjust your windows so that you can see both Blender and the video.,estelle,,,
-BLENDMONSTER2_INFO_3,Here's the link - https://www.youtube.com/watch?v=L0AY61v6-M4,estelle,,,
+BLENDMONSTER2_INFO_3,"Here's the link - <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://www.youtube.com/watch?v=L0AY61v6-M4</span></u>",estelle,,,
 BLENDMONSTER2_INFO_4,"Have fun! Remember, you can always restart this activity if you need to see the tips or link.",estelle,,,
