+++ data/quests_strings/blendmonster4.csv
@@ -2,12 +2,12 @@ BLENDMONSTER4_QUEST_NAME,Monster Smoothie - Texturing your Surfaces,,,,
 BLENDMONSTER4_QUEST_SUBTITLE,What you'll do,,,,
 BLENDMONSTER4_QUEST_DESCRIPTION,"In part 4 of this quest, you'll explore materials and surface textures. Give your characters a creepy look!",,,,
 BLENDMONSTER4_QUIT,See you later!,,,,
-BLENDMONSTER4_NOTINSTALLED,"Uh-oh! Looks like `Blender` isn't installed!
+BLENDMONSTER4_NOTINSTALLED,"Uh-oh! Looks like <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Blender</span></tt> isn't installed!
 
 You need to download it from the App Center if you'd like to learn how to use it.
 
-Click the icon named `App Center` on the Desktop, then search for `Blender`, and install it. ",daemon,,,
-BLENDMONSTER4_INFO_1,"Welcome back, *{{user_name}}*! I'd recommend doing the previous parts of this quest if you haven't already. If you have, or you're forging ahead, plug in your mouse and prepare for Materials!",estelle,,,
+Click the icon named <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">App Center</span></tt> on the Desktop, then search for <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Blender</span></tt>, and install it. ",daemon,,,
+BLENDMONSTER4_INFO_1,"Welcome back, <b>{{user_name}}</b>! I'd recommend doing the previous parts of this quest if you haven't already. If you have, or you're forging ahead, plug in your mouse and prepare for Materials!",estelle,,,
 BLENDMONSTER4_INFO_2,The easiest way to follow along with these tutorials is to adjust your windows so that you can see both Blender and the video.,estelle,,,
-BLENDMONSTER4_INFO_3,Here's the link - https://www.youtube.com/watch?v=OFmKedu8r88,estelle,,,
+BLENDMONSTER4_INFO_3,"Here's the link - <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://www.youtube.com/watch?v=OFmKedu8r88</span></u>",estelle,,,
 BLENDMONSTER4_INFO_4,"Have fun! Remember, you can always restart this activity if you need to see the tips or link.",estelle,,,
