+++ data/quests_strings/blendmonster7.csv
@@ -2,7 +2,7 @@ BLENDMONSTER7_QUEST_NAME,Monster Smoothie - The Files,,,,
 BLENDMONSTER7_QUEST_SUBTITLE,What you'll do,,,,
 BLENDMONSTER7_QUEST_DESCRIPTION,"In this supplement to the Monster Smoothie quests, you'll get access to the files used to make the videos you learned from.",,,,
 BLENDMONSTER7_QUIT,See you later!,,,,
-BLENDMONSTER7_INFO_1,"Hi, *{{user_name}}*. Did you enjoy the *Monster Smoothie* activity? Even if you didn't get all the way though, I think you'll be interested in this.",estelle,,,
-BLENDMONSTER7_INFO_2,All the files that were used to make the videos you learned from are now in your *home* directory - `~/BlenderSources/MonsterSmoothieFiles`,estelle,,,
+BLENDMONSTER7_INFO_1,"Hi, <b>{{user_name}}</b>. Did you enjoy the <b>Monster Smoothie</b> activity? Even if you didn't get all the way though, I think you'll be interested in this.",estelle,,,
+BLENDMONSTER7_INFO_2,"All the files that were used to make the videos you learned from are now in your <b>home</b> directory - <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">~/BlenderSources/MonsterSmoothieFiles</span></tt>",estelle,,,
 BLENDMONSTER7_INFO_3,"I hope you'll find these useful - try and find any differences between how you did something, and how the originals did it. You might find some interesting tricks!",estelle,,,
 BLENDMONSTER7_INFO_4,I hope you had lots of fun making your monster. See you again soon!,estelle,,,
