+++ data/quests_strings/htmlintro1.csv
@@ -1,22 +1,22 @@
 HTMLINTRO1_QUEST_NAME,Introduction to HTML,,,,
 HTMLINTRO1_QUEST_SUBTITLE,What you'll do,,,,
 HTMLINTRO1_QUEST_DESCRIPTION,Learn the basics of web pages and the language they're written in  - HTML. (Part 1),,,,
-HTMLINTRO1_1,"Hi, *{{user_name}}*! I know you've used the internet before - but have you ever wondered what goes into showing you all those web pages?",riley,,,
+HTMLINTRO1_1,"Hi, <b>{{user_name}}</b>! I know you've used the internet before - but have you ever wondered what goes into showing you all those web pages?",riley,,,
 HTMLINTRO1_2,"I was wondering that, one day, and I decided to go find out. There was a ton to learn, but now that I've done the legwork, I can pass it on to you.",riley,,,
 HTMLINTRO1_3,"Before I explain anything, let's get you looking at the guts of a web page.",riley,,,
 HTMLINTRO1_4,"The page should open for you, but if it doesn't, try running your web browser.",riley,,,
-HTMLINTRO1_5,You can also try clicking this link: https://codepen.io/madetohack/pen/BaaNeXj?editors=1000#code-area,riley,,,
+HTMLINTRO1_5,"You can also try clicking this link: <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://codepen.io/madetohack/pen/BaaNeXj?editors=1000</span></u>#code-area",riley,,,
 HTMLINTRO1_6,OK! Here we are. One half of your browser should be showing you the code that the page is written in.,riley,,,
 HTMLINTRO1_7,The other half shows what you'd see if you found this page on the internet.,riley,,,
-HTMLINTRO1_8,"If you want the split to work a different way, you can resize your browser's window to show the *Change View* button on the upper right part of this webpage.",riley,,,
-HTMLINTRO1_9,"Once you can see that button, you can click it, and change the *Editor Layout* to a different one.",riley,,,
+HTMLINTRO1_8,"If you want the split to work a different way, you can resize your browser's window to show the <b>Change View</b> button on the upper right part of this webpage.",riley,,,
+HTMLINTRO1_9,"Once you can see that button, you can click it, and change the <b>Editor Layout</b> to a different one.",riley,,,
 HTMLINTRO1_10,"Just like computer programs, web pages are written in a kind of code - It's called HTML.",riley,,,
 HTMLINTRO1_11,"You put HTML around text, and that tells your web browser what to show, and how to show it on your screen. Not so bad, right?",riley,,,
-HTMLINTRO1_12,"In HTML, everything with `<` and `>` around it is called a `tag`. Tags tell the browser how to display something - text, images, whatever.",riley,,,
-HTMLINTRO1_13,"Tags usually come in pairs - an `<opening tag>`, and a `</closing tag>`, like saying *start* and *stop*.",riley,,,
-HTMLINTRO1_14,"You'll also see some text with `<!--` and `-->` around it. Those are called *comments*, it's info I've put there to help explain what some of the HTML does.",,,,
-HTMLINTRO1_15,"Let's get our hands dirty! Find the text *Hello, World!* in the code part of this page, and change it - make it your name, your favorite food, whatever you like.",riley,,,
+HTMLINTRO1_12,"In HTML, everything with <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;</span></tt> and <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&gt;</span></tt> around it is called a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">tag</span></tt>. Tags tell the browser how to display something - text, images, whatever.",riley,,,
+HTMLINTRO1_13,"Tags usually come in pairs - an <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;opening tag&gt;</span></tt>, and a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;/closing tag&gt;</span></tt>, like saying <b>start</b> and <b>stop</b>.",riley,,,
+HTMLINTRO1_14,"You'll also see some text with <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;!--</span></tt> and <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">--&gt;</span></tt> around it. Those are called <b>comments</b>, it's info I've put there to help explain what some of the HTML does.",,,,
+HTMLINTRO1_15,"Let's get our hands dirty! Find the text <b>Hello, World!</b> in the code part of this page, and change it - make it your name, your favorite food, whatever you like.",riley,,,
 HTMLINTRO1_16,"Awesome! You just edited a webpage! That makes you a web designer, and that's pretty cool.",riley,,,
-HTMLINTRO1_17,Let's take it one step further. Do you see the *Welcome* text at the top of the page? Try finding that in the code and changing it.,riley,,,
-HTMLINTRO1_18,That's right - it's between the `<h1>` tags. Nice job! ,riley,,,
-HTMLINTRO1_19,"See, HTML isn't as confusing as you thought, right? Find me back in the Clubhouse, and we'll continue the adventure in a more... *colorful*... way!",riley,,,
+HTMLINTRO1_17,Let's take it one step further. Do you see the <b>Welcome</b> text at the top of the page? Try finding that in the code and changing it.,riley,,,
+HTMLINTRO1_18,"That's right - it's between the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;h1&gt;</span></tt> tags. Nice job! ",riley,,,
+HTMLINTRO1_19,"See, HTML isn't as confusing as you thought, right? Find me back in the Clubhouse, and we'll continue the adventure in a more... <b>colorful</b>... way!",riley,,,
