+++ data/quests_strings/htmlintro2.csv
@@ -1,30 +1,30 @@
 HTMLINTRO2_QUEST_NAME,HTML - Using Colors,,,,
 HTMLINTRO2_QUEST_SUBTITLE,What you'll do,,,,
 HTMLINTRO2_QUEST_DESCRIPTION,Take your HTML knowledge to the next level by learning about web colors. (Part 2),,,,
-HTMLINTRO2_1,"Hey again! Ready to dive a little deeper, *{{user_name}}*?",riley,,,
+HTMLINTRO2_1,"Hey again! Ready to dive a little deeper, <b>{{user_name}}</b>?",riley,,,
 HTMLINTRO2_2,First things first - let's get back to that web page previewer.,riley,,,
-HTMLINTRO2_3,"If it's not opening, run or switch to your browser. You can hold *Alt* and press *Tab* to cycle through active programs, or look at the bottom of the screen.",riley,,,
-HTMLINTRO2_4,You can also click this link: https://codepen.io/madetohack/pen/BaaKggj?editors=1000,riley,,,
+HTMLINTRO2_3,"If it's not opening, run or switch to your browser. You can hold <b>Alt</b> and press <b>Tab</b> to cycle through active programs, or look at the bottom of the screen.",riley,,,
+HTMLINTRO2_4,"You can also click this link: <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://codepen.io/madetohack/pen/BaaKggj?editors=1000</span></u>",riley,,,
 HTMLINTRO2_5,"Here's a fun fact - Just based on our last lesson, you already know enough to make your own webpage!",riley,,,
 HTMLINTRO2_6,"If you did, though, all you'd have is boring black text on a white background. Let's spice that up a little!",riley,,,
-HTMLINTRO2_7,"Find an opening `<p>` tag in your code, and add `style=”color:blue”` after the `p`. It should look like this: `<p style=""color:blue"">`",riley,,,
-HTMLINTRO2_8,Congratulations! You just *styled* some text! Now let's get even fancier.,riley,,,
-HTMLINTRO2_9,Find another `<p>` tag and make *that* text orange.,riley,,,
-HTMLINTRO2_10,"If you did that right, the tag should look like this: `<p style=""color:orange"">`, and your text should be orange!",riley,,,
-HTMLINTRO2_11,I wonder what happens if you try the same thing with an `<h1>` or `<h2>` tag? Try making one of those purple.,riley,,,
-HTMLINTRO2_12,"If you did it right, it should look something like `<h1 style=""color:purple"">`",riley,,,
+HTMLINTRO2_7,"Find an opening <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p&gt;</span></tt> tag in your code, and add <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">style=”color:blue”</span></tt> after the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">p</span></tt>. It should look like this: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p style=""color:blue""&gt;</span></tt>",riley,,,
+HTMLINTRO2_8,Congratulations! You just <b>styled</b> some text! Now let's get even fancier.,riley,,,
+HTMLINTRO2_9,"Find another <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p&gt;</span></tt> tag and make <b>that</b> text orange.",riley,,,
+HTMLINTRO2_10,"If you did that right, the tag should look like this: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p style=""color:orange""&gt;</span></tt>, and your text should be orange!",riley,,,
+HTMLINTRO2_11,"I wonder what happens if you try the same thing with an <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;h1&gt;</span></tt> or <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;h2&gt;</span></tt> tag? Try making one of those purple.",riley,,,
+HTMLINTRO2_12,"If you did it right, it should look something like <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;h1 style=""color:purple""&gt;</span></tt>",riley,,,
 HTMLINTRO2_13,"Nice. What if you wanted to use a super-special shade of green, though? How would you do that?",riley,,,
 HTMLINTRO2_14,"It'd be really cool if we had a more specific way to express color, where we could say exactly how we want to mix it.",riley,,,
-HTMLINTRO2_15,"You see where I'm going with this, right? We totally do have that! It's called *hex color*, and it lets you create colors by using 6-digit numbers.",riley,,,
-HTMLINTRO2_16,Here's an example hex color - `#76EECF`,riley,,,
-HTMLINTRO2_17,"To use this hex color in your webpage, replace one of your color words with that hex code: `#76EECF`",riley,,,
-HTMLINTRO2_18,"If you edited a `<p>` tag, it would be `<p style=""color:#76EECF"">`",riley,,,
+HTMLINTRO2_15,"You see where I'm going with this, right? We totally do have that! It's called <b>hex color</b>, and it lets you create colors by using 6-digit numbers.",riley,,,
+HTMLINTRO2_16,"Here's an example hex color - <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">#76EECF</span></tt>",riley,,,
+HTMLINTRO2_17,"To use this hex color in your webpage, replace one of your color words with that hex code: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">#76EECF</span></tt>",riley,,,
+HTMLINTRO2_18,"If you edited a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p&gt;</span></tt> tag, it would be <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p style=""color:#76EECF""&gt;</span></tt>",riley,,,
 HTMLINTRO2_19,Cool! Change the numbers around and see if you can create another color!,riley,,,
 HTMLINTRO2_20,"Hex color uses 1 - 9, and A B C D E F as numbers - It's a little out of our way right now to explain why, I can tell you later if you're interested.",riley,,,
-HTMLINTRO2_21,"If you want to use a color picker, you can try this webpage: https://htmlcolorcodes.com/",riley,,,
-HTMLINTRO2_22,"Ok, so you've got some color chops now. But do you know what would *really* make this page look cool?",riley,,,
-HTMLINTRO2_23,I think we need to change the background color to something less boring. Look for the `<body>` tag in the code.,riley,,,
-HTMLINTRO2_24,"It should be near the top of the page. Inside that tag, put `style=""background-color:coral""`, just like you put the `color` information into a `<p>` tag.",riley,,,
-HTMLINTRO2_25,"Your body tag should look like this -`<body style=""background-color:coral"">`",riley,,,
+HTMLINTRO2_21,"If you want to use a color picker, you can try this webpage: <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://htmlcolorcodes.com/</span></u>",riley,,,
+HTMLINTRO2_22,"Ok, so you've got some color chops now. But do you know what would <b>really</b> make this page look cool?",riley,,,
+HTMLINTRO2_23,"I think we need to change the background color to something less boring. Look for the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;body&gt;</span></tt> tag in the code.",riley,,,
+HTMLINTRO2_24,"It should be near the top of the page. Inside that tag, put <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">style=""background-color:coral""</span></tt>, just like you put the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">color</span></tt> information into a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p&gt;</span></tt> tag.",riley,,,
+HTMLINTRO2_25,"Your body tag should look like this -<tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;body style=""background-color:coral""&gt;</span></tt>",riley,,,
 HTMLINTRO2_26,Niiice! Now that's going to turn some heads.,riley,,,
 HTMLINTRO2_27,"So, how does it feel to have the tools to paint the whole internet? Come find me back in the Clubhouse to keep going!",riley,,,
