+++ data/quests_strings/htmlintro4.csv
@@ -1,28 +1,28 @@
 HTMLINTRO4_QUEST_NAME,"HTML - CSS, The Secret Sauce",,,,
 HTMLINTRO4_QUEST_SUBTITLE,What you'll do,,,,
 HTMLINTRO4_QUEST_DESCRIPTION,"Learn the basics of CSS, the ""secret sauce"" that makes styling web pages a snap! (Part 4)",,,,
-HTMLINTRO4_1,Good to see you! I hope you're ready to learn some of the *inner secrets* of web design ^_^,riley,,,
+HTMLINTRO4_1,Good to see you! I hope you're ready to learn some of the <b>inner secrets</b> of web design ^_^,riley,,,
 HTMLINTRO4_2,"As always - if the previewer page doesn't open, try opening your browser.",riley,,,
-HTMLINTRO4_3,Or you can click the link - https://codepen.io/madetohack/pen/vYYyNba?editors=1000,riley,,,
-HTMLINTRO4_4,"Well, here it is - there's a way to automate all those `color=` and `font-size=` bits you've been typing out one at a time.",,,,
-HTMLINTRO4_5,It's called CSS - that stands for *cascading style sheets* - and it's pretty awesome. ,,,,
+HTMLINTRO4_3,"Or you can click the link - <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://codepen.io/madetohack/pen/vYYyNba?editors=1000</span></u>",riley,,,
+HTMLINTRO4_4,"Well, here it is - there's a way to automate all those <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">color=</span></tt> and <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">font-size=</span></tt> bits you've been typing out one at a time.",,,,
+HTMLINTRO4_5,It's called CSS - that stands for <b>cascading style sheets</b> - and it's pretty awesome. ,,,,
 HTMLINTRO4_6,"Up until now, we've been using mostly HTML, which controls the content of pages. CSS lets us really dig into how the content is shown to everyone who visits the page.",riley,,,
 HTMLINTRO4_7,"Think of it like food - HTML is rice and chicken, and the CSS is all the different recipes that use those two basic building blocks. The food is the same, but the spices, colors, flavors... they're all different.",,,,
-HTMLINTRO4_8,"You should notice a brand-new tag inside the `<head>` tag, `<style>`",riley,,,
-HTMLINTRO4_9,"Remember how to change text colors? You modified a `p` or `h1` tag with `style=""color:green""`.",riley,,,
-HTMLINTRO4_10,"The `<style>` tag lets you centralize all that manual work, by creating a set of rules that can apply to all the tags in your webpage, like ""Make paragraphs purple"".",riley,,,
-HTMLINTRO4_11,"Adding rules to the `<style>` tags tells the browser ""when you see this tag, apply these changes to it.""",riley,,,
-HTMLINTRO4_12,"Let's give it a shot. To start, let's make all the paragraphs purple. Between `<style>` and `</style>`, type `p { color:purple; }`",riley,,,
-HTMLINTRO4_13,"That's telling the browser ""every time you see the `<p>` tag, add `color=""purple""` to it.""",,,,
-HTMLINTRO4_14,"At this point, you should have `<style> p { color:purple; } </style>` (but on 3 separate lines)",,,,
+HTMLINTRO4_8,"You should notice a brand-new tag inside the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;head&gt;</span></tt> tag, <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt>",riley,,,
+HTMLINTRO4_9,"Remember how to change text colors? You modified a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">p</span></tt> or <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">h1</span></tt> tag with <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">style=""color:green""</span></tt>.",riley,,,
+HTMLINTRO4_10,"The <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt> tag lets you centralize all that manual work, by creating a set of rules that can apply to all the tags in your webpage, like ""Make paragraphs purple"".",riley,,,
+HTMLINTRO4_11,"Adding rules to the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt> tags tells the browser ""when you see this tag, apply these changes to it.""",riley,,,
+HTMLINTRO4_12,"Let's give it a shot. To start, let's make all the paragraphs purple. Between <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt> and <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;/style&gt;</span></tt>, type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">p { color:purple; }</span></tt>",riley,,,
+HTMLINTRO4_13,"That's telling the browser ""every time you see the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;p&gt;</span></tt> tag, add <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">color=""purple""</span></tt> to it.""",,,,
+HTMLINTRO4_14,"At this point, you should have <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt; p { color:purple; } &lt;/style&gt;</span></tt> (but on 3 separate lines)",,,,
 HTMLINTRO4_15,"That's pretty stylish! Now that we've got some kickin' purple text, let's make it bigger.",riley,,,
-HTMLINTRO4_16,"Add `font-size: x-large;` right after the color rule, keeping it inside the braces ( *{ }* ) Don't forget a space between the rules, or the semicolon (*;*) after each one.",riley,,,
-HTMLINTRO4_17,Your `<style>` tag should now have `p { color:purple; font-size: x-large; }` inside it.,riley,,,
-HTMLINTRO4_18,Check out that text! Now *that's* got my attention.,riley,,,
-HTMLINTRO4_19,"You can style other tags like this, too. Use the name of the tag, an open brace ( *{* ), your styling, and then a closing brace ( *}* ).",riley,,,
-HTMLINTRO4_20,Let's try making your `h1` text red. ,riley,,,
-HTMLINTRO4_21,"To do that, you'd put `h1 { color:red; }` on the line below your instructions for `p`.",riley,,,
+HTMLINTRO4_16,"Add <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">font-size: x-large;</span></tt> right after the color rule, keeping it inside the braces ( <b>{ }</b> ) Don't forget a space between the rules, or the semicolon (<b>;</b>) after each one.",riley,,,
+HTMLINTRO4_17,"Your <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt> tag should now have <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">p { color:purple; font-size: x-large; }</span></tt> inside it.",riley,,,
+HTMLINTRO4_18,Check out that text! Now <b>that's</b> got my attention.,riley,,,
+HTMLINTRO4_19,"You can style other tags like this, too. Use the name of the tag, an open brace ( <b>{</b> ), your styling, and then a closing brace ( <b>}</b> ).",riley,,,
+HTMLINTRO4_20,"Let's try making your <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">h1</span></tt> text red. ",riley,,,
+HTMLINTRO4_21,"To do that, you'd put <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">h1 { color:red; }</span></tt> on the line below your instructions for <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">p</span></tt>.",riley,,,
 HTMLINTRO4_22,"Great job! Do you see how easy this can be, now?",riley,,,
-HTMLINTRO4_23,"It doesn't matter how much text or how many tags you have in that page, the `<style>` tag will apply to all of it.",riley,,,
+HTMLINTRO4_23,"It doesn't matter how much text or how many tags you have in that page, the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;style&gt;</span></tt> tag will apply to all of it.",riley,,,
 HTMLINTRO4_24,"There's still another level beyond this, though - What if I told you that you could make up your own tags?",riley,,,
 HTMLINTRO4_25,Come find me in the Clubhouse soon if you want to learn it!,riley,,,
