+++ data/quests_strings/inspector1.csv
@@ -1,21 +1,21 @@
 INSPECTOR1_QUEST_NAME,An Unexpected Inspection,,,,
 INSPECTOR1_QUEST_SUBTITLE,What you'll do,,,,
 INSPECTOR1_QUEST_DESCRIPTION,"Ada reveals a tool for hacking webpages, hidden inside your browser! (Part 1)",,,,
-INSPECTOR1_1,"Hello, *{{user_name}}*. I hope you've been having fun with Riley, learning about the web.",ada,,,
+INSPECTOR1_1,"Hello, <b>{{user_name}}</b>. I hope you've been having fun with Riley, learning about the web.",ada,,,
 INSPECTOR1_2,"Learning to program is like learning a language - you can learn to write it, but you also need to know how to read. You might be able to create a webpage, but can you decode one in its natural habitat?",ada,,,
-INSPECTOR1_3,Let's go take a look at a real webpage and see if we can pick it apart. Take us to `http://hack-computer.com/about`,ada,,,
+INSPECTOR1_3,"Let's go take a look at a real webpage and see if we can pick it apart. Take us to <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF""><u><span insert_hyphens=""false"" foreground=""#3584E4"">http://hack-computer.com/about</span></u></span></tt>",ada,,,
 INSPECTOR1_4,"Here we are! Normally, when you see a webpage, what you're looking at is the result of your browser reading all the code in the webpage, and then interpreting it.",ada,,,
-INSPECTOR1_5,"If we want to see the actual code, we need to use the *page Inspector*. This tool will let you view and change the code of any webpage!",ada,,,
-INSPECTOR1_6,Let's open the Inspector. Right-click (or two-finger click on the touchpad if you're using a Hack Laptop) anywhere on the page and select *Inspect*. You can also just press `Ctrl + Shift + I`,ada,,,
-INSPECTOR1_7,"Oh my, that's quite a lot of complicated code! Don't worry, though, I'll walk you through it. First of all, you should notice that the Inpector popped in from the *right-hand* side of the browser.",ada,,,
+INSPECTOR1_5,"If we want to see the actual code, we need to use the <b>page Inspector</b>. This tool will let you view and change the code of any webpage!",ada,,,
+INSPECTOR1_6,"Let's open the Inspector. Right-click (or two-finger click on the touchpad if you're using a Hack Laptop) anywhere on the page and select <b>Inspect</b>. You can also just press <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Ctrl + Shift + I</span></tt>",ada,,,
+INSPECTOR1_7,"Oh my, that's quite a lot of complicated code! Don't worry, though, I'll walk you through it. First of all, you should notice that the Inpector popped in from the <b>right-hand</b> side of the browser.",ada,,,
 INSPECTOR1_8,"You can drag the border between the page and the Inspector to adjust how much screen space the Inspector takes up, if you need more room.",ada,,,
-INSPECTOR1_9,"You should also see that the Inspector is divided into two halves. For now, we're only interested in the half that says *Elements* at the top. (You can grab the bar between these two halves and drag it to give youself more space if you like.)",ada,,,
-INSPECTOR1_10,"Now, click the icon in the *top left corner of the Inspector* - it should look like an arrow pointing at a box. This is the *Select Element* button - a fancy way of saying ""it's for picking the thing you want to look at.""",ada,,,
+INSPECTOR1_9,"You should also see that the Inspector is divided into two halves. For now, we're only interested in the half that says <b>Elements</b> at the top. (You can grab the bar between these two halves and drag it to give youself more space if you like.)",ada,,,
+INSPECTOR1_10,"Now, click the icon in the <b>top left corner of the Inspector</b> - it should look like an arrow pointing at a box. This is the <b>Select Element</b> button - a fancy way of saying ""it's for picking the thing you want to look at.""",ada,,,
 INSPECTOR1_11,"Now, mouse over the left-hand side of your browser window, where the actual webpage is. Notice how different parts of the page turn different colors? If you look back to the Inspector, you'll see the code for the colored area get highlighted as you move the mouse.",ada,,,
-INSPECTOR1_12,"You can move the mouse as much as you like to highlight elements, but when you click, it will deactivate the *Select Element* button. You'll have to turn it back on again if you want to select another element.",ada,,,
-INSPECTOR1_13,Click the words “About Us” on the webpage (not the Inspector). Notice that the Inspector will highlight the line that reads: `<span style=”font-family...”> Hack Laptop </span>.`,ada,,,
-INSPECTOR1_14,"In the Inspector pane, within the highlighted line, double-click the word “About”. This will select the text, and let you edit it. Change it to something else, then hit `Enter`!",ada,,,
-INSPECTOR1_15,"Now, look back at the left-hand pane. *About Us* should be gone, and your text should be in its place! Now *that's* hacking a webpage!",ada,,,
+INSPECTOR1_12,"You can move the mouse as much as you like to highlight elements, but when you click, it will deactivate the <b>Select Element</b> button. You'll have to turn it back on again if you want to select another element.",ada,,,
+INSPECTOR1_13,"Click the words “About Us” on the webpage (not the Inspector). Notice that the Inspector will highlight the line that reads: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;span style=”font-family...”&gt; Hack Laptop &lt;/span&gt;.</span></tt>",ada,,,
+INSPECTOR1_14,"In the Inspector pane, within the highlighted line, double-click the word “About”. This will select the text, and let you edit it. Change it to something else, then hit <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Enter</span></tt>!",ada,,,
+INSPECTOR1_15,"Now, look back at the left-hand pane. <b>About Us</b> should be gone, and your text should be in its place! Now <b>that's</b> hacking a webpage!",ada,,,
 INSPECTOR1_16,"At this point, you might be wondering if anyone else saw the changes we made. No, they only happened on your computer. When you load a webpage in your browser, you're getting a copy of the webpage as it is on the server. If you refresh, all your changes will disappear - you're getting a new copy.",ada,,,
 INSPECTOR1_17,"Now that you know about the Inspector, you can pick apart and hack any webpage you come across! You can even use this to play pranks on your friends...",ada,,,
 INSPECTOR1_18,"That's enough for the moment, but I've got some more tricks ready for the Inspector. Come back for part 2 and I'll show you!",ada,,,
