+++ data/quests_strings/makerearthquake.csv
@@ -1,5 +1,5 @@
 MAKEREARTHQUAKE_QUEST_NAME,A Home-grown Earthquake,,,,
-MAKEREARTHQUAKE_WELCOME,"Now this is some mad-scientist tech! Ever wanted to control the *mighty power of earthquakes*?
+MAKEREARTHQUAKE_WELCOME,"Now this is some mad-scientist tech! Ever wanted to control the <b>mighty power of earthquakes</b>?
 
 Who hasn't? I think you'll have a lot of fun with this one!",faber,,,
 MAKEREARTHQUAKE_QUEST_SUBTITLE,What you'll do,,,,
