+++ data/quests_strings/migration.csv
@@ -1,32 +1,32 @@
 MIGRATION_QUEST_NAME,,,,,
-MIGRATION_WELCOME,"Hey, *{{user_name}}*! Welcome back to Hack! There have been quite a few changes around here recently, so I thought I'd run you through them.",ada,,,
+MIGRATION_WELCOME,"Hey, <b>{{user_name}}</b>! Welcome back to Hack! There have been quite a few changes around here recently, so I thought I'd run you through them.",ada,,,
 MIGRATION_WELCOME2,"The Clubhouse is no longer on the side of the screen - Now it's an icon on the Desktop, like any other app.",ada,,,
 MIGRATION_OPENIT,Give it a click!,ada,,,
 MIGRATION_OLDSTUFF1,"Do you like the new place? We've just moved in, but it already feels like home.",ada,,,
 MIGRATION_OLDSTUFF2,"If you're wondering how to get to the games you used to play in the old Clubhouse, don't worry.",ada,,,
 MIGRATION_OLDSTUFF3,"We upgraded your permissions to the maximum level - everything's unlocked for you, now!",ada,,,
-MIGRATION_OLDSTUFF4,"You can go anywhere you like in *Sidetrack*, *Lightspeed*, and *Fizzics*.",ada,,,
+MIGRATION_OLDSTUFF4,"You can go anywhere you like in <b>Sidetrack</b>, <b>Lightspeed</b>, and <b>Fizzics</b>.",ada,,,
 MIGRATION_OLDSTUFF5,"It's one of our ways of saying ""thanks"" to you for sticking with us in our early days.",ada,,,
 MIGRATION_NEWSTUFF,"Now, let's talk about the new Clubhouse.",ada,,,
-MIGRATION_HACKMODE1,"There's a new system we've added, called *Hack Mode*. When you're in Hack Mode (like you are now), you can use flip-to-hack on all the apps and games that support it.",ada,,,
+MIGRATION_HACKMODE1,"There's a new system we've added, called <b>Hack Mode</b>. When you're in Hack Mode (like you are now), you can use flip-to-hack on all the apps and games that support it.",ada,,,
 MIGRATION_HACKMODE2,"Also, your desktop and cursor will change to remind you.",ada,,,
-MIGRATION_HACKMODE3,That *big switch on the right* controls Hack Mode. Give it a try!,ada,,,
+MIGRATION_HACKMODE3,That <b>big switch on the right</b> controls Hack Mode. Give it a try!,ada,,,
 MIGRATION_HACKMODE4,"If you switch Hack Mode OFF, you’ll go back to the regular desktop, and all your hacking tools will shut down.",ada,,,
 MIGRATION_HACKMODE5,"Think of it sort of like 'stealth mode' - you can use it if someone else needs to use the computer, or you need to do your homework.",ada,,,
 MIGRATION_ACTIVITIES1,"In this new version of Hack, you can play whatever activities you want, whenever you want.",ada,,,
-MIGRATION_ACTIVITIES2,"*Up top*, you'll see the Pathways button - Pathways organize activities into categories like Web, OS, Games, Art, and Maker.",ada,,,
+MIGRATION_ACTIVITIES2,"<b>Up top</b>, you'll see the Pathways button - Pathways organize activities into categories like Web, OS, Games, Art, and Maker.",ada,,,
 MIGRATION_ACTIVITIES3,You can also talk to all of us - we've got the activities organized by our common interests.,ada,,,
-MIGRATION_PROFILE1,Do you see that picture with your name on it in the *upper left*? That's your *Profile*.,ada,,,
-MIGRATION_PROFILE2,"Everyone who uses Hack is different, and we want to recognize that. Did you notice how I knew your name earlier, *{{user_name}}*?",ada,,,
+MIGRATION_PROFILE1,Do you see that picture with your name on it in the <b>upper left</b>? That's your <b>Profile</b>.,ada,,,
+MIGRATION_PROFILE2,"Everyone who uses Hack is different, and we want to recognize that. Did you notice how I knew your name earlier, <b>{{user_name}}</b>?",ada,,,
 MIGRATION_PROFILE3,"Now, if you want to be called something else, you can change that.",ada,,,
 MIGRATION_PROFILE_ASK,Would you like to learn how to do that right now?,ada,,,
 MIGRATION_PROFILE_POS,Sure!,ada,,,
 MIGRATION_PROFILE_NEG,"Not now, thanks.",ada,,,
-MIGRATION_PROFILE_CHANGENAME1,"If you look in the *far bottom right* corner of the whole screen, you'll see a little picture.
+MIGRATION_PROFILE_CHANGENAME1,"If you look in the <b>far bottom right</b> corner of the whole screen, you'll see a little picture.
 
-Click that, and it'll pop up a menu. From that menu, click *Settings*",ada,,,
-MIGRATION_PROFILE_CHANGENAME2,"Once you've clicked that, you'll be in the *Settings* app.",ada,,,
-MIGRATION_PROFILE_CHANGENAME3,"Go to the *Details* category (you might need to scroll down), and then the *Users* category.",ada,,,
+Click that, and it'll pop up a menu. From that menu, click <b>Settings</b>",ada,,,
+MIGRATION_PROFILE_CHANGENAME2,"Once you've clicked that, you'll be in the <b>Settings</b> app.",ada,,,
+MIGRATION_PROFILE_CHANGENAME3,"Go to the <b>Details</b> category (you might need to scroll down), and then the <b>Users</b> category.",ada,,,
 MIGRATION_PROFILE_CHANGENAME4,All done? Let's move on.,ada,,,
 MIGRATION_END1,"That's pretty much all the major changes to Hack - it's a lot, I know, but we think it's a big improvement. Explore the Clubhouse and check out all the new things there are to do!",ada,,,
 MIGRATION_END2,And from all of us here - thanks for coming back and sticking with us. We've got a lot of cool stuff planned!,ada,,,
