+++ data/quests_strings/osoneshotbackground.csv
@@ -4,12 +4,12 @@ OSONESHOTBACKGROUND_QUEST_DESCRIPTION,Learn how to change your desktop backgroun
 OSONESHOTBACKGROUND_LAUNCH,"Run the Terminal, and let's get to it.",riley,,,
 OSONESHOTBACKGROUND_1,"What a ride! We decoded a hidden message, followed a treasure map, and walked away with pirate's booty!",riley,,,
 OSONESHOTBACKGROUND_2,This wouldn't be complete without a way to show off your prize - and I've done a little digging to figure out how.,riley,,,
-OSONESHOTBACKGROUND_3,"Let's get to where we stored your prize - `cd Pictures`, then `ls` to make sure the file's there.",riley,,,
+OSONESHOTBACKGROUND_3,"Let's get to where we stored your prize - <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cd Pictures</span></tt>, then <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">ls</span></tt> to make sure the file's there.",riley,,,
 OSONESHOTBACKGROUND_4,"Great! Let's set this image as your desktop background, now.",riley,hurray,,
-OSONESHOTBACKGROUND_5,"Look down in the bottom right of the screen, and click your user picture. From that popup menu, click `Settings`.",riley,,,
-OSONESHOTBACKGROUND_6,The `Settings` app is where all the internals of the OS are controlled. Look on the left side of the window and click `Background`.,riley,,,
-OSONESHOTBACKGROUND_7,"... and then click `Add Picture...` in the upper right. The file browser will pop up, and you can find `treasure.jpg` that way.",riley,,,
+OSONESHOTBACKGROUND_5,"Look down in the bottom right of the screen, and click your user picture. From that popup menu, click <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Settings</span></tt>.",riley,,,
+OSONESHOTBACKGROUND_6,"The <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Settings</span></tt> app is where all the internals of the OS are controlled. Look on the left side of the window and click <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Background</span></tt>.",riley,,,
+OSONESHOTBACKGROUND_7,"... and then click <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">Add Picture...</span></tt> in the upper right. The file browser will pop up, and you can find <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">treasure.jpg</span></tt> that way.",riley,,,
 OSONESHOTBACKGROUND_8,"Remember, no matter how you look at the file system, it's all the same thing. The Terminal and a browser like this are all looking at the same data.",riley,,,
 OSONESHOTBACKGROUND_9,"Once you find the image, double-click it - that'll add it to the list in the Settings window. Then, click that image in the list, and select ""Set Background"".",riley,,,
 OSONESHOTBACKGROUND_10,"And there you go! If anyone wants to know where you got that cool background, send 'em to me!",riley,,,
-OSONESHOTBACKGROUND_11,"Thanks for letting me teach you some cool stuff, *{{user_name}}* - I had a lot of fun doing it. Keep an eye out for more!",riley,hurray,,
+OSONESHOTBACKGROUND_11,"Thanks for letting me teach you some cool stuff, <b>{{user_name}}</b> - I had a lot of fun doing it. Keep an eye out for more!",riley,hurray,,
