+++ data/quests_strings/osoneshotfollow.csv
@@ -3,25 +3,25 @@ OSONESHOTFOLLOW_QUEST_SUBTITLE,What you'll do,,,,
 OSONESHOTFOLLOW_QUEST_DESCRIPTION,Learn to follow a map through your computer's file system and move files. (Part 3 of 4),,,,
 OSONESHOTFOLLOW_LAUNCH,"Run the Terminal from the desktop, and let's roll out!",riley,,,
 OSONESHOTFOLLOW_1,"I'm so excited - We're on the trail of something cool, I know it. Let's take another look at that map.",riley,,,
-OSONESHOTFOLLOW_2,"If you're not in `yarnbasket`, let's get back in there, type `cd yarnbasket`.",riley,,,
-OSONESHOTFOLLOW_3,"Let's look at your results from last time - if you forgot the file you saved, type `ls` and try to find it.",riley,,,
-OSONESHOTFOLLOW_4,Now you can `cat results` (or whatever you named your file).,riley,,,
+OSONESHOTFOLLOW_2,"If you're not in <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">yarnbasket</span></tt>, let's get back in there, type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cd yarnbasket</span></tt>.",riley,,,
+OSONESHOTFOLLOW_3,"Let's look at your results from last time - if you forgot the file you saved, type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">ls</span></tt> and try to find it.",riley,,,
+OSONESHOTFOLLOW_4,"Now you can <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cat results</span></tt> (or whatever you named your file).",riley,,,
 OSONESHOTFOLLOW_5,"Like I said last time, this looks like a list of directories. Let's follow the trail!",riley,,,
-OSONESHOTFOLLOW_6,"I'll let you in on a cool trick you can use to travel through lots of directories at once. Normally, you'd just type `cd first`, `cd second`, and so on. But!",riley,,,
-OSONESHOTFOLLOW_7,There's a faster way. You can type `cd all/those/directories/in/a/row`. Give it a shot!,riley,,,
+OSONESHOTFOLLOW_6,"I'll let you in on a cool trick you can use to travel through lots of directories at once. Normally, you'd just type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cd first</span></tt>, <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cd second</span></tt>, and so on. But!",riley,,,
+OSONESHOTFOLLOW_7,"There's a faster way. You can type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cd all/those/directories/in/a/row</span></tt>. Give it a shot!",riley,,,
 OSONESHOTFOLLOW_8,Is... that a message in the directory names? Wierd! I wonder what it means?,riley,,,
-OSONESHOTFOLLOW_9,"Now that you're at the end, let's `ls`",riley,,,
-OSONESHOTFOLLOW_10,Is that a file named `treasure`? Haha! We found it!,riley,hurray,,
+OSONESHOTFOLLOW_9,"Now that you're at the end, let's <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">ls</span></tt>",riley,,,
+OSONESHOTFOLLOW_10,"Is that a file named <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">treasure</span></tt>? Haha! We found it!",riley,hurray,,
 OSONESHOTFOLLOW_11,"I feel like a pirate digging up buried treasure. Yarr, matey! Avast!",riley,hurray,,
-OSONESHOTFOLLOW_12,Let's investigate! `cat treasure.jpg`!,riley,,,
-OSONESHOTFOLLOW_13,"Oh, right. I feel a little dumb right now - this is a `jpg` file, which is an image. You can't get anything out of one of these by using `cat`.",riley,,,
+OSONESHOTFOLLOW_12,"Let's investigate! <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cat treasure.jpg</span></tt>!",riley,,,
+OSONESHOTFOLLOW_13,"Oh, right. I feel a little dumb right now - this is a <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">jpg</span></tt> file, which is an image. You can't get anything out of one of these by using <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cat</span></tt>.",riley,,,
 OSONESHOTFOLLOW_14,"Well, let's move this file to a place where we can look at it more easily - like your home directory.",riley,,,
-OSONESHOTFOLLOW_15,"To move a file, we need to use the *move* command... `mv`. ",riley,,,
-OSONESHOTFOLLOW_16,"To use `mv` to move this file, it needs the file you want to move, and where you want it to go.",riley,,,
-OSONESHOTFOLLOW_17,"So, that's `mv treasure.jpg ∼/Pictures/`",riley,,,
-OSONESHOTFOLLOW_18,"Nice. Now the file should be in your *Pictures* folder, inside your home directory.",riley,,,
-OSONESHOTFOLLOW_19,"If you want, you can open the file manager program (*Documents* on your Desktop) and go see that the file is there.",riley,,,
+OSONESHOTFOLLOW_15,"To move a file, we need to use the <b>move</b> command... <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">mv</span></tt>. ",riley,,,
+OSONESHOTFOLLOW_16,"To use <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">mv</span></tt> to move this file, it needs the file you want to move, and where you want it to go.",riley,,,
+OSONESHOTFOLLOW_17,"So, that's <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">mv treasure.jpg ∼/Pictures/</span></tt>",riley,,,
+OSONESHOTFOLLOW_18,"Nice. Now the file should be in your <b>Pictures</b> folder, inside your home directory.",riley,,,
+OSONESHOTFOLLOW_19,"If you want, you can open the file manager program (<b>Documents</b> on your Desktop) and go see that the file is there.",riley,,,
 OSONESHOTFOLLOW_20,"BTW, if you're wondering how I know all these commands and how to use them, well, it's because I've got a reference.",riley,,,
-OSONESHOTFOLLOW_21,"I actually wrote it myself! If you want to see it, I dropped it in your *Documents* folder - it's called `rileysreference.txt`",riley,,,
-OSONESHOTFOLLOW_22,"Remember, if you need to read a text file in the Terminal, you can use `cat`!",riley,,,
+OSONESHOTFOLLOW_21,"I actually wrote it myself! If you want to see it, I dropped it in your <b>Documents</b> folder - it's called <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">rileysreference.txt</span></tt>",riley,,,
+OSONESHOTFOLLOW_22,"Remember, if you need to read a text file in the Terminal, you can use <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">cat</span></tt>!",riley,,,
 OSONESHOTFOLLOW_23,Whew. I think that's enough for the moment! I have something special for your final adventure - check back with me soon!,riley,,,
