+++ data/quests_strings/p5jsintro2.csv
@@ -1,26 +1,26 @@
 P5JSINTRO2_QUEST_NAME,P5JS - Making it interactive!,,,,
 P5JSINTRO2_QUEST_SUBTITLE,What you'll do,,,,
 P5JSINTRO2_QUEST_DESCRIPTION,Take your P5JS knowledge to the next level by making it interactive. (Part 2),,,,
-P5JSINTRO2_1,"Hi there! I'm glad you're back, *{{user_name}}* - I've got something awesome to show you!",riley,,,
-P5JSINTRO2_2,"First things first - let's get back to that code editor. If you closed it after last time, you can start fresh at https://editor.p5js.org",riley,,,
-P5JSINTRO2_3,"Last time we made the background prettier using the code `background('green')` or `background('#76EECF')`, and then we drew a circle on top of it.",riley,,,
-P5JSINTRO2_4,"The circle code was `circle(100,200,50)` giving it an X-coordinate of 100, a Y-coordinate of 200, and a diameter of 50.",riley,,,
-P5JSINTRO2_5,"In order to make the circle a color other than white, we have to use a command called `fill()`",riley,,,
-P5JSINTRO2_6,Go ahead and make a new line of code before the `circle()` command that says `fill('blue');` and run the code by clicking play (or auto-refresh).,riley,,,
+P5JSINTRO2_1,"Hi there! I'm glad you're back, <b>{{user_name}}</b> - I've got something awesome to show you!",riley,,,
+P5JSINTRO2_2,"First things first - let's get back to that code editor. If you closed it after last time, you can start fresh at <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://editor.p5js.org</span></u>",riley,,,
+P5JSINTRO2_3,"Last time we made the background prettier using the code <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">background('green')</span></tt> or <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">background('#76EECF')</span></tt>, and then we drew a circle on top of it.",riley,,,
+P5JSINTRO2_4,"The circle code was <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">circle(100,200,50)</span></tt> giving it an X-coordinate of 100, a Y-coordinate of 200, and a diameter of 50.",riley,,,
+P5JSINTRO2_5,"In order to make the circle a color other than white, we have to use a command called <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">fill()</span></tt>",riley,,,
+P5JSINTRO2_6,"Go ahead and make a new line of code before the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">circle()</span></tt> command that says <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">fill('blue');</span></tt> and run the code by clicking play (or auto-refresh).",riley,,,
 P5JSINTRO2_7,"Ah, that's better. But what about that annoying black line around the outside? ",riley,,,
-P5JSINTRO2_8,"That's called the *stroke*, and we can get rid of it by typing `strokeWeight(0);` on a new line.",riley,,,
-P5JSINTRO2_9,"Of course, we could also make it thicker by typing *strokeWeight(10)* instead - the number in parentheses is how many pixels wide you want it.",riley,,,
-P5JSINTRO2_10,And we can even change the color of the outline - try `stroke('red');` or `stroke('#E84904');`,riley,,,
-P5JSINTRO2_11,"Once you have your circle looking like you want it, let's take another look at our code that draws the circle: `circle(100,200,50)`",riley,,,
-P5JSINTRO2_12,"Remember how I said the first number in parentheses was the *X-coordinate*? X means side-to-side, or how far left or right the circle is.",riley,,,
-P5JSINTRO2_13,"What if instead of drawing that circle at X-coordinate *100* every time, we drew the circle at the X-coordinate of my cursor?",riley,,,
-P5JSINTRO2_14,"To do this, replace the number *100* with *mouseX* and run the program.",riley,,,
-P5JSINTRO2_15,"You should have `circle(mouseX, 200, 50);` and if you move your mouse back and forth along the canvas, you should see it following along left to right!",riley,,,
+P5JSINTRO2_8,"That's called the <b>stroke</b>, and we can get rid of it by typing <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">strokeWeight(0);</span></tt> on a new line.",riley,,,
+P5JSINTRO2_9,"Of course, we could also make it thicker by typing <b>strokeWeight(10)</b> instead - the number in parentheses is how many pixels wide you want it.",riley,,,
+P5JSINTRO2_10,"And we can even change the color of the outline - try <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">stroke('red');</span></tt> or <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">stroke('#E84904');</span></tt>",riley,,,
+P5JSINTRO2_11,"Once you have your circle looking like you want it, let's take another look at our code that draws the circle: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">circle(100,200,50)</span></tt>",riley,,,
+P5JSINTRO2_12,"Remember how I said the first number in parentheses was the <b>X-coordinate</b>? X means side-to-side, or how far left or right the circle is.",riley,,,
+P5JSINTRO2_13,"What if instead of drawing that circle at X-coordinate <b>100</b> every time, we drew the circle at the X-coordinate of my cursor?",riley,,,
+P5JSINTRO2_14,"To do this, replace the number <b>100</b> with <b>mouseX</b> and run the program.",riley,,,
+P5JSINTRO2_15,"You should have <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">circle(mouseX, 200, 50);</span></tt> and if you move your mouse back and forth along the canvas, you should see it following along left to right!",riley,,,
 P5JSINTRO2_16,"Notice that the circle is not moving up and down, because its Y-coordinate is still 200 every time. But we can change that, too!",riley,,,
-P5JSINTRO2_17,"If you type `circle(mouseX, mouseY, 50);`, you can make the circle follow your mouse exactly throughout the canvas!",riley,,,
+P5JSINTRO2_17,"If you type <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">circle(mouseX, mouseY, 50);</span></tt>, you can make the circle follow your mouse exactly throughout the canvas!",riley,,,
 P5JSINTRO2_18,"Wheeeeee! Sorry, I'm having too much fun...",riley,,,
 P5JSINTRO2_19,"I have more to show you, but before we go too much farther, let's learn how to save our work.",riley,,,
-P5JSINTRO2_20,"At the top of the screen, there is a *File* menu. Click on it, and you will find the option to *Save*.",riley,,,
+P5JSINTRO2_20,"At the top of the screen, there is a <b>File</b> menu. Click on it, and you will find the option to <b>Save</b>.",riley,,,
 P5JSINTRO2_21,"Note that in order to save your work, you have to sign up with an email address and make an account.",riley,,,
 P5JSINTRO2_22,"If you don't have an email address or you're not allowed to sign up, no worries! Just copy and paste your code somewhere for safekeeping.",riley,,,
 P5JSINTRO2_23,"Once your code is saved, you can close your computer, go stretch your legs, and come back when you're ready to learn the next steps!",riley,,,
