+++ data/quests_strings/p5jsintro3.csv
@@ -2,16 +2,16 @@ P5JSINTRO3_QUEST_NAME,"P5JS - ""Painting"" on the canvas",,,,
 P5JSINTRO3_QUEST_SUBTITLE,What you'll do,,,,
 P5JSINTRO3_QUEST_DESCRIPTION,"Learn some key differences between the setup() and draw() functions that allow you to ""paint""! (Part 3)",,,,
 P5JSINTRO3_1,Welcome back! I'm excited to show you what I figured out - this is gonna be fun!,riley,,,
-P5JSINTRO3_2,"First things first - let's get back to that code editor. If you closed it after last time, go to https://editor.p5js.org and log in to get it back.",riley,,,
+P5JSINTRO3_2,"First things first - let's get back to that code editor. If you closed it after last time, go to <u><span insert_hyphens=""false"" foreground=""#3584E4"">https://editor.p5js.org</span></u> and log in to get it back.",riley,,,
 P5JSINTRO3_3,"You should have a circle on the screen that moves around to follow your mouse. If not, you can go back to the last quest to get caught up.",riley,,,
-P5JSINTRO3_4,Now it's time to take a closer look at the two main functions in this program: `setup()` and `draw()`,riley,,,
-P5JSINTRO3_5,"The `setup()` function happens once, at the very beginning. Currently, it just sets up the canvas and tells it what size to be.",riley,,,
-P5JSINTRO3_6,"The `draw()` function happens over and over again, as long as the program is running. Every few milliseconds, it runs the `draw()` code again.",riley,,,
+P5JSINTRO3_4,"Now it's time to take a closer look at the two main functions in this program: <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">setup()</span></tt> and <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">draw()</span></tt>",riley,,,
+P5JSINTRO3_5,"The <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">setup()</span></tt> function happens once, at the very beginning. Currently, it just sets up the canvas and tells it what size to be.",riley,,,
+P5JSINTRO3_6,"The <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">draw()</span></tt> function happens over and over again, as long as the program is running. Every few milliseconds, it runs the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">draw()</span></tt> code again.",riley,,,
 P5JSINTRO3_7,That's why it looks like the circle is moving with our mouse. The program redraws it at our current mouseX and mouseY every few milliseconds!,riley,,,
 P5JSINTRO3_8,"It also redraws the background each time, covering up the last circle we drew.",riley,,,
 P5JSINTRO3_9,"But what would happen if we didn't redraw the background each time? What if we only drew it once, at the very beginning?",riley,,,
-P5JSINTRO3_10,"Let's find out! If we move the `background()` line of code to the `setup()` function, it will only happen once, right after the canvas is created.",riley,,,
-P5JSINTRO3_11,Try it - move the `background();` line to right after the `createCanvas();` line. Make sure you put it before the `}` so it's in the right function.,riley,,,
+P5JSINTRO3_10,"Let's find out! If we move the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">background()</span></tt> line of code to the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">setup()</span></tt> function, it will only happen once, right after the canvas is created.",riley,,,
+P5JSINTRO3_11,"Try it - move the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">background();</span></tt> line to right after the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">createCanvas();</span></tt> line. Make sure you put it before the <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">}</span></tt> so it's in the right function.",riley,,,
 P5JSINTRO3_12,Now run the program - what happens?,riley,,,
 P5JSINTRO3_13,"Woah! Is that what you thought would happen? If you want to start over, you can just hit the play button again.",riley,,,
 P5JSINTRO3_14,So... did you notice how there's always a circle at the top left corner when you start? Why do you think that is?,riley,,,
@@ -19,6 +19,6 @@ P5JSINTRO3_15,Let's think about it - where is your mouse when the program starts
 P5JSINTRO3_16,"Exactly - your mouse is outside the canvas, so it draws the circle as close as it can. ",riley,,,
 P5JSINTRO3_17,Try weaving your mouse in and out of the canvas - you can decorate all the corners without touching the middle!,riley,,,
 P5JSINTRO3_18,Can you figure out how to draw a happy face?,riley,,,
-P5JSINTRO3_19,"If you still have an outline around your circle, try using `strokeWeight(0)` to get rid of it and have a true paintbrush.",riley,,,
-P5JSINTRO3_20,"Woohoo! What else can you paint? To take a screenshot of your creation, press the 'print screen' key on your keyboard - the image will show up in your *Pictures* folder.",riley,,,
-P5JSINTRO3_21,I'lll see you next time - we're going to keep digging deeper into *p5.js* and see what we can find!,riley,,,
+P5JSINTRO3_19,"If you still have an outline around your circle, try using <tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">strokeWeight(0)</span></tt> to get rid of it and have a true paintbrush.",riley,,,
+P5JSINTRO3_20,"Woohoo! What else can you paint? To take a screenshot of your creation, press the 'print screen' key on your keyboard - the image will show up in your <b>Pictures</b> folder.",riley,,,
+P5JSINTRO3_21,I'lll see you next time - we're going to keep digging deeper into <b>p5.js</b> and see what we can find!,riley,,,
