+++ data/quests_strings/quickstart.csv
@@ -1,19 +1,19 @@
 QUICKSTART_NAME,Tutorial - The Clubhouse - Quickstart,,,,
-QUICKSTART_WELCOME1,"Hello, *{{user_name}}*! I'm Ada - and let me be the first to welcome you to Hack!",ada,,,
+QUICKSTART_WELCOME1,"Hello, <b>{{user_name}}</b>! I'm Ada - and let me be the first to welcome you to Hack!",ada,,,
 QUICKSTART_WELCOME2,"I'm going to run you through how this program, The Clubhouse, works. Is that OK?",ada,,,
 QUICKSTART_POSITIVE,👍,ada,,,
 QUICKSTART_NEGATIVE,👎,ada,,,
-QUICKSTART_HACKSWITCH1,This is the main area of the Clubhouse. You can always use the *left-most top button* to get back here.,ada,,,
-QUICKSTART_HACKSWITCH2,"That big switch on the *right* controls *Hack Mode*. Right now, it's *ON*, so you have access to everything in the Clubhouse. Try it out!",ada,,,
+QUICKSTART_HACKSWITCH1,This is the main area of the Clubhouse. You can always use the <b>left-most top button</b> to get back here.,ada,,,
+QUICKSTART_HACKSWITCH2,"That big switch on the <b>right</b> controls <b>Hack Mode</b>. Right now, it's <b>ON</b>, so you have access to everything in the Clubhouse. Try it out!",ada,,,
 QUICKSTART_HACKSWITCH3,"If the switch is off, the Clubhouse turns off, and you're essentially using a normal computer.",ada,,,
 QUICKSTART_ACTIVITIES,"To get your adventure started, talk to any one of us in the Clubhouse. We've all got different interests, so there's a lot of different things to do.",ada,,,
-QUICKSTART_PATHWAYS1,"The *middle top button* is the Pathways button. If you already know exactly what kind of activities you want to do, go there.",ada,,,
+QUICKSTART_PATHWAYS1,"The <b>middle top button</b> is the Pathways button. If you already know exactly what kind of activities you want to do, go there.",ada,,,
 QUICKSTART_PATHWAYS2,"If you prefer to explore, talk to everyone here in the Clubhouse - we've got lots of interesting activities to do.",ada,,,
-QUICKSTART_PROFILE1,Your profile is in the upper left. You can change your name and profile picture using the Endless OS *Settings*.,ada,,,
+QUICKSTART_PROFILE1,Your profile is in the upper left. You can change your name and profile picture using the Endless OS <b>Settings</b>.,ada,,,
 QUICKSTART_PROFILE2,"To get directly to that section, click the picture way down in the far bottom right of the whole screen.
 
-Then, in that popup menu, click *Settings*.
+Then, in that popup menu, click <b>Settings</b>.
 
-From there, you can go to the *Details* category (you might need to scroll down), then the *Users* category.",ada,,,
+From there, you can go to the <b>Details</b> category (you might need to scroll down), then the <b>Users</b> category.",ada,,,
 QUICKSTART_END1,"OK, that covers the basics! Come and find me if you want more detail. Have fun!",ada,hurray,,
 QUICKSTART_DECLINE,"Can't wait? I know the feeling! If you need any help, just click on me in the Clubhouse, and I'll fill you in.",ada,,,
