+++ data/quests_strings/story_ada1.csv
@@ -3,13 +3,13 @@ STORY_ADA1_QUEST_SUBTITLE,In this story:,,,,
 STORY_ADA1_QUEST_DESCRIPTION,Learn about the purpose of the Academy and get to know Ada a little better.,,,,
 STORY_ADA1_POSITIVE,👍,,,,
 STORY_ADA1_NEGATIVE,👎,,,,
-STORY_ADA1_1,"Hi again, *{{user_name}}*! I was so excited when Dr. Rowe mentioned that the remote student program was approved.",ada,,,
+STORY_ADA1_1,"Hi again, <b>{{user_name}}</b>! I was so excited when Dr. Rowe mentioned that the remote student program was approved.",ada,,,
 STORY_ADA1_2,"I know it's not easy - you're the first person to test this system, and you have to deal with a totally new set of people, in a new place.",ada,,,
 STORY_ADA1_3,"So, how are you doing?",ada,,,
-STORY_ADA1_3_RPOS,"Wow, you *are* a fast learner. I knew you'd be a good fit here - It's great to see you succeed so easily.",ada,hurray,,
+STORY_ADA1_3_RPOS,"Wow, you <b>are</b> a fast learner. I knew you'd be a good fit here - It's great to see you succeed so easily.",ada,hurray,,
 STORY_ADA1_3_RNEG,"Don't get discouraged - I know it's hard to be so limited in what you can say and do. If you keep going, I'm sure you'll figure it out.",ada,,,
 STORY_ADA1_4,All of us at the Academy are here to support you. We want you to be able to learn whatever you want! ,ada,,,
-STORY_ADA1_5,"One of our goals is to let you hack *any* program in the operating system. Don't worry, we'll teach you how to do it when the time comes.",ada,,,
+STORY_ADA1_5,"One of our goals is to let you hack <b>any</b> program in the operating system. Don't worry, we'll teach you how to do it when the time comes.",ada,,,
 STORY_ADA1_6,"Your arrival at The Academy is the start of an exciting journey - We want everyone to learn about computers, programming, and the digital world.",ada,,,
 STORY_ADA1_7,"All over the world, people's lives are becoming more and more digital. If you're not connected, you could be left behind.",ada,,,
 STORY_ADA1_8,"It hasn't been too long since you had to use a computer to use the internet, and cell phones could only make phone calls.",ada,,,
@@ -17,11 +17,11 @@ STORY_ADA1_9,"Now, every cell phone can use the Internet, plus hundreds of diffe
 STORY_ADA1_10,"All this technology creates new problems, though. The people that control a program or a service get to decide how other people can use it.",ada,,,
 STORY_ADA1_11,"For example - If you control a search engine, you control what information people can find with it.",ada,,,
 STORY_ADA1_12,"If you own a search engine, and you don't like dogs, you can make it so that nobody can ever find information about dogs. That's a big problem!",ada,,,
-STORY_ADA1_13,"This is why you need to learn how these systems work. If you know how something works, you can make it work *for you*, instead of as a way to control you.",ada,,,
+STORY_ADA1_13,"This is why you need to learn how these systems work. If you know how something works, you can make it work <b>for you</b>, instead of as a way to control you.",ada,,,
 STORY_ADA1_14,"Knowing the system means that other people will have a harder time controlling what you can learn, do, or say. That's why learning about all this is so important!",ada,,,
-STORY_ADA1_15,"… Wow, that was quite a speech. Sorry, these things are *REALLY* important to me -  I get all excited to talk about them and just *go*.",ada,,,
+STORY_ADA1_15,"… Wow, that was quite a speech. Sorry, these things are <b>REALLY</b> important to me -  I get all excited to talk about them and just <b>go</b>.",ada,,,
 STORY_ADA1_16,"Hey Ada, how's it going?",faber,,,
-STORY_ADA1_17,"Hi, Faber! I was just seeing how *{{user_name}}* was adjusting to the Academy.",ada,,,
+STORY_ADA1_17,"Hi, Faber! I was just seeing how <b>{{user_name}}</b> was adjusting to the Academy.",ada,,,
 STORY_ADA1_18,"Cool! I actually had some data I wanted to collect - I'm trying to optimize the program that shows your avatar to everyone at The Academy. Hopefully I can keep you from walking through solid objects, and other things like that.",faber,,,
 STORY_ADA1_19,"It's funny if you know what to expect, but man, if you popped out of a wall in front of me I'd yell loud enough to wake Dr. Rowe out of a nap!",faber,,,
 STORY_ADA1_20,"Haha! I just had the image of you jumping, and everything in your toolbelt flying all over…",ada,,,
@@ -34,7 +34,7 @@ STORY_ADA1_26,"Yeah, she's the best, even if she does scare the living daylights
 STORY_ADA1_27,"It's moments like this that make me so glad I'm here, at the Academy.",ada,,,
 STORY_ADA1_28,"If I had left, I'd probably be making some useless gadget for people who don't need more distractions. What about you?",faber,,,
 STORY_ADA1_29,"Oh, I imagine I'd be at some research center. Maybe working on gravity waves. This is much more rewarding, though.",ada,,,
-STORY_ADA1_30,"So, *{{user_name}}*, how are you feeling about the Academy so far?",faber,,,
+STORY_ADA1_30,"So, <b>{{user_name}}</b>, how are you feeling about the Academy so far?",faber,,,
 STORY_ADA1_30_RPOS,Awesome! I love seeing new people get all charged up about this place - it's such a cool opportunity.,faber,hurray,,
 STORY_ADA1_30_RNEG,"Don't worry - Just learn as much as you can. I didn't know what I wanted to do until this one night, when I realized I had been in the machine shop since 10AM that morning... and it was 4AM, the next morning.",faber,,,
 STORY_ADA1_31,What was that noise?,ada,,,
@@ -43,4 +43,4 @@ Great talking with you guys, but I gotta run and load this into my program if I 
 STORY_ADA1_33,In time for what?,ada,,,
 STORY_ADA1_34,"Dr. Hart got me a spot on the servers so I can run my simulations... But there's a lot of activity on the servers these days, so I've got a pretty tight deadline.",faber,,,
 STORY_ADA1_35,"Good luck, Faber!
-Well, *{{user_name}}*, I have to be going - but I'll see you later. Have fun!",ada,,,
+Well, <b>{{user_name}}</b>, I have to be going - but I'll see you later. Have fun!",ada,,,
