+++ data/quests_strings/story_faber1.csv
@@ -3,7 +3,7 @@ STORY_FABER1_QUEST_SUBTITLE,In this story:,,,,
 STORY_FABER1_QUEST_DESCRIPTION,"Meet Faber Mattics, Maker Space coordinator and legendary laid-back dude.",,,,
 STORY_FABER1_POSITIVE,👍,,,,
 STORY_FABER1_NEGATIVE,👎,,,,
-STORY_FABER1_1,"Hey there, *{{user_name}}* - I don't know if we've been fully introduced. I’m Faber, I run the Maker Space out near the Observatory.",faber,,,
+STORY_FABER1_1,"Hey there, <b>{{user_name}}</b> - I don't know if we've been fully introduced. I’m Faber, I run the Maker Space out near the Observatory.",faber,,,
 STORY_FABER1_2,"It’s always great to have another new student around - Say, do you like shortbread?",faber,,,
 STORY_FABER1_2_RPOS,Awesome! I’ll double my recipe next time I make some. Can’t wait for you to try it.,faber,,,
 STORY_FABER1_2_RNEG,"That’s cool - Let me know what your favorite cookie is, we can try a batch in my solar oven some time!",faber,,,
@@ -16,7 +16,7 @@ STORY_FABER1_8,"I'm sure there were cookies involved, too. You've got a clever w
 STORY_FABER1_9,"Hey, just trying to make the trip worth it for everyone. You know, if all you wanted was to bring more people out, you could help me design a transport system…",faber,,,
 STORY_FABER1_10,"Knowing you, Faber, you'd be trying to get me to build some kind of rocket-powered train.",estelle,,,
 STORY_FABER1_11,"Hey, they say to use what you know, and you know rockets! I've seen your welds before - superb.",faber,,,
-STORY_FABER1_12,"It *is* always nice to have your work complimented, but these days I investigate the stars mainly through a telescope.",estelle,,,
+STORY_FABER1_12,"It <b>is</b> always nice to have your work complimented, but these days I investigate the stars mainly through a telescope.",estelle,,,
 STORY_FABER1_13,"And your simulations, of course - what's the latest?",faber,,,
 STORY_FABER1_14,"It's still running - I launched a probe on a long-track slingshot out beyond the heliopause - it should only take a week or so more, I think.",estelle,,,
 STORY_FABER1_15,That's past the edge of the solar system! Got any theories about what you'll find?,faber,,,
