+++ data/quests_strings/story_riley1.csv
@@ -7,13 +7,13 @@ STORY_RILEY1_1,"Hi! I knew we were getting a new student at the Academy, as part
 STORY_RILEY1_2,"I'm Riley, and I am sooo excited to have you here! I was part of some advanced program-thing, and there's nobody my age around, so…",riley,,,
 STORY_RILEY1_3,"Anyway! You're here! I have to show you SO MANY things, this place is amazing.",riley,,,
 STORY_RILEY1_4,"I, uh, don't want to pry, but… what's up with your avatar? It's like you're not really here, or something.",riley,,,
-STORY_RILEY1_5,"There you are, Riley! I was hoping you two would find each other! Riley, *{{user_name}}* is our remote student. They're on a very low-bandwidth link- remember your networking lessons?",ada,,,
+STORY_RILEY1_5,"There you are, Riley! I was hoping you two would find each other! Riley, <b>{{user_name}}</b> is our remote student. They're on a very low-bandwidth link- remember your networking lessons?",ada,,,
 STORY_RILEY1_6,"Oh, so they can't answer me?",riley,,,
-STORY_RILEY1_7,"They can, but only with 1 bit - a 'yes' or a 'no' only. Isn't that right, *{{user_name}}*?",ada,,,
+STORY_RILEY1_7,"They can, but only with 1 bit - a 'yes' or a 'no' only. Isn't that right, <b>{{user_name}}</b>?",ada,,,
 STORY_RILEY1_7_RPOS,See?,ada,,,
 STORY_RILEY1_7_RNEG,Your sense of humor certainly matches Riley's. I see some joint pranks in your future…,ada,,,
 STORY_RILEY1_8,"It's ok you can't say much  - Saniel says I talk more than enough on my own! Actually, he says:",riley,,,
-STORY_RILEY1_9,"*""Riley, have you been altering the acoustics of my classroom again? Every time I detect several voices, they're all yours. That's quite a trick… but this is History of Algorithms, not Ventriloquism.""*",riley,,,
+STORY_RILEY1_9,"<b>""Riley, have you been altering the acoustics of my classroom again? Every time I detect several voices, they're all yours. That's quite a trick… but this is History of Algorithms, not Ventriloquism.""</b>",riley,,,
 STORY_RILEY1_10,That's quite a good imitation of Dr. Rowe.,ada,,,
 STORY_RILEY1_11,"It is, isn't it? I'm glad to hear that at least some part of me seems to be rubbing off on her.",saniel,,,
 STORY_RILEY1_12,"Saniel! Uh, hi.",riley,,,
@@ -24,9 +24,9 @@ STORY_RILEY1_16,And just when you think you'll never find her and stop…,ada,,,
 STORY_RILEY1_17,"I know, right? She'll pop up right next to you! I want to learn how to do THAT.",riley,,,
 STORY_RILEY1_18,"You'd no doubt spend the first week or so trying to see how far you could make us jump in fright. Then again, you would always be on time to class, I suppose. Perhaps it's worth the trade?",saniel,,,
 STORY_RILEY1_19,"Aw, come on, that was only one week!",riley,,,
-STORY_RILEY1_20,"Don't worry, Riley, Dr. Rowe still won't let me use the book-shelving robot after *the incident*.",ada,,,
+STORY_RILEY1_20,"Don't worry, Riley, Dr. Rowe still won't let me use the book-shelving robot after <b>the incident</b>.",ada,,,
 STORY_RILEY1_21,Just because there's a 'high' speed setting doesn't mean that it's safe. You nearly took my head off!,saniel,,,
-STORY_RILEY1_22,"I stopped it at *least* 10 feet from you, and one of the cats jumped onto the controls!",ada,,,
+STORY_RILEY1_22,"I stopped it at <b>least</b> 10 feet from you, and one of the cats jumped onto the controls!",ada,,,
 STORY_RILEY1_23,"I suppose a high-speed trip around the archives isn't the worst thing that could happen to a cat. It didn't take too long to clean up, either.",saniel,,,
 STORY_RILEY1_24,"Hmm. Speaking of cats... Riley, could you come help me feed the feline masses, please?",saniel,,,
-STORY_RILEY1_25,"Sure thing, Dr. Rowe! I gotta run, *{{user_name}}*, but it was great to meet you! I can't wait to see the fun things we'll do in the future.",riley,,,
+STORY_RILEY1_25,"Sure thing, Dr. Rowe! I gotta run, <b>{{user_name}}</b>, but it was great to meet you! I can't wait to see the fun things we'll do in the future.",riley,,,
