+++ data/quests_strings/story_riley2.csv
@@ -5,43 +5,43 @@ STORY_RILEY2_POSITIVE,👍,,,,
 STORY_RILEY2_NEGATIVE,👎,,,,
 STORY_RILEY2_1,Hey! I realized that I forgot to introduce you to one of the most important residents of the Academy when we first met.,riley,,,
 STORY_RILEY2_2,"It's OK, though, since he was hiding anyway. May I present to you the Felicitous Ephemeral Living Intelligence, eXperimental… Felix!",riley,,,
-STORY_RILEY2_3,_Mrrow._,felixnet,,,
+STORY_RILEY2_3,<i>Mrrow.</i>,felixnet,,,
 STORY_RILEY2_4,Felix. Buddy. Help me out here. Do a trick? Please?,riley,,,
-STORY_RILEY2_5,`<Felix licks his paw.>`,felixnet,,,
+STORY_RILEY2_5,"<tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;Felix licks his paw.&gt;</span></tt>",felixnet,,,
 STORY_RILEY2_6,Anything?,riley,,,
-STORY_RILEY2_7,`<Felix stares off into the distance.>`,felixnet,,,
+STORY_RILEY2_7,"<tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;Felix stares off into the distance.&gt;</span></tt>",felixnet,,,
 STORY_RILEY2_8,"OK, so his name doesn't stand for anything, it's just Felix, but he is really impressive from at least one point of view -",riley,,,
 STORY_RILEY2_9,"I've never seen a program act so convincingly like a cat. Especially, you know… the annoying parts.
 Like not listening to anything I ask him to do! Ever!",riley,,,
-STORY_RILEY2_10,_Puurrrrow?_,felixnet,,,
+STORY_RILEY2_10,<i>Puurrrrow?</i>,felixnet,,,
 STORY_RILEY2_11,"Also being incredibly cute! I mean, seriously, look at those paws. And that face! Actually don't look too long, I think there's some kind of fractal thing going on there, and it starts to move — ",riley,,,
-STORY_RILEY2_12,"— But someone was clearly trying to capture the *idea* of a cat, and I think they nailed it.",riley,,,
+STORY_RILEY2_12,"— But someone was clearly trying to capture the <b>idea</b> of a cat, and I think they nailed it.",riley,,,
 STORY_RILEY2_13,"Felix has been around the Academy as long as anyone can remember, but he’s been out and about much more often since I arrived.
 Ada says it must be my natural charm that lured him out.",riley,,,
 STORY_RILEY2_14,"Actually, Riley, I think I said ""your prank-loving nature,"" but natural charm is probably just as likely.",ada,,,
-STORY_RILEY2_15,Hey Ada! I was just introducing Felix. He seems really interested in *{{user_name}}*'s avatar!,riley,,,
-STORY_RILEY2_16,"Look at that, he's… sniffing? Can he sniff even if he doesn't have a nose? *{{user_name}}*, we’d better take a look at your connection.
+STORY_RILEY2_15,Hey Ada! I was just introducing Felix. He seems really interested in <b>{{user_name}}</b>'s avatar!,riley,,,
+STORY_RILEY2_16,"Look at that, he's… sniffing? Can he sniff even if he doesn't have a nose? <b>{{user_name}}</b>, we’d better take a look at your connection.
 Riley, do you have your network analyzer with you?",ada,,,
 STORY_RILEY2_17,"Always! Let's see… open Wireshark… There we go, that's the secure tunnel… and yeah, wow. Look at all that network traffic!",riley,,,
-STORY_RILEY2_18,"Interesting. He seems to be looking at all the data that's connected to your avatar, *{{user_name}}*. What's he so interested in?",ada,,,
+STORY_RILEY2_18,"Interesting. He seems to be looking at all the data that's connected to your avatar, <b>{{user_name}}</b>. What's he so interested in?",ada,,,
 STORY_RILEY2_19,"Maybe he's just… doing cat stuff. You know, poking around and trying to see everything?",riley,,,
 STORY_RILEY2_20,"It very well could be that. Felix's programming could just be directing him to look at everything he can find. If that were true, though, then we'd see him all over, all the time. He'd be quite a distraction during exams.",ada,,,
 STORY_RILEY2_21,"Aw, did you have to remind me about exams?",riley,,,
 STORY_RILEY2_22,"That's months away, Riley.",ada,,,
 STORY_RILEY2_23,"Yeah, but it's still a drag! You know I'm all about my classes, why do I have to do some big test?",riley,,,
 STORY_RILEY2_24,"Testing is an important part of learning - It's about being able to apply your knowledge to a new problem, with restrictions. It's the surest way of seeing whether you really learned the lessons or not.",ada,,,
-STORY_RILEY2_25,"Hmm. New problems… maybe that's why Felix is so interested. You're the newest thing here, *{{user_name}}*! Everything you do is new to Felix!",ada,,,
+STORY_RILEY2_25,"Hmm. New problems… maybe that's why Felix is so interested. You're the newest thing here, <b>{{user_name}}</b>! Everything you do is new to Felix!",ada,,,
 STORY_RILEY2_26,"So is he… learning? If he's learning, does that mean…",riley,,,
 STORY_RILEY2_27,"I don't think he's an AI, Riley. If he's showing responses that you're never seen before, it might be due to new events.
-*""When you hear hoofbeats, think horses, not Zebras.""*",ada,,,
+<b>""When you hear hoofbeats, think horses, not Zebras.""</b>",ada,,,
 STORY_RILEY2_28,"So... that means… he's a horse? Seriously, what does that mean?",riley,,,
 STORY_RILEY2_29,"It means that you should always consider the most commonly found solution to a problem first, before moving to less common possibilities.",ada,,,
 STORY_RILEY2_30,"For example: All the words in one of your papers turning into ""meow"" just two hours before the deadline... well, I think you see where I'm going with this.",ada,,,
 STORY_RILEY2_31,"Hey, come on, what else would a cat type?",riley,,,
 STORY_RILEY2_32,"I did let you turn that paper in late, Riley. Dr. Rowe might not have been quite as understanding. And while we're on the subject-",ada,,,
-STORY_RILEY2_33,_Brrrip! Mrzzrzrzrzrrhhhh…_,felixnet,,,
+STORY_RILEY2_33,<i>Brrrip! Mrzzrzrzrzrrhhhh…</i>,felixnet,,,
 STORY_RILEY2_34,"Felix! Don't corrupt my papers, it's not nice.",riley,,,
-STORY_RILEY2_35,_Nyan._,felixnet,,,
+STORY_RILEY2_35,<i>Nyan.</i>,felixnet,,,
 STORY_RILEY2_36,"Wait, what's that in his mouth?",riley,,,
 STORY_RILEY2_37,Is that…?,ada,,,
 STORY_RILEY2_38,"False alarm! Looks like it's a fragment of code that shows up… as a mouse. Yeech. Gross, but impressive, all at the same time.",riley,,,
@@ -51,10 +51,10 @@ STORY_RILEY2_41,"Hmm. Analyzing - File header is consistent with source code, ru
 STORY_RILEY2_42,"Uh, a little lost here…",riley,,,
 STORY_RILEY2_43,"Oh! I'm sorry, I got carried away. I think it's code, but very tightly packed, to the point where it's barely readable by a human.",ada,,,
 STORY_RILEY2_44,"It's very clever, but not clever enough to hide everything it's doing... Calls to network functions... uploads to multiple different servers... Felix, where did you find this?",ada,,,
-STORY_RILEY2_45,"_Brbrbrbr._
-`<Felix gurgles smugly.>`",felixnet,,,
+STORY_RILEY2_45,"<i>Brbrbrbr.</i>
+<tt><span insert_hyphens=""false"" foreground=""#287A8C"" background=""#FFFFFF"">&lt;Felix gurgles smugly.&gt;</span></tt>",felixnet,,,
 STORY_RILEY2_46,Does he do that just mess with us?,riley,,,
 STORY_RILEY2_47,"I'm sure he's just proud that he's gotten us to pay so much attention to him. By the way, can I have a copy of… the ""mouse""?",ada,,,
 STORY_RILEY2_48,"Go for it! I'm going to try and figure out how that file was using a fake image… Maybe I can use it bait my ""Felix trap"" - or, as I call it, 'infinite-ear-scratches.exe'",riley,,,
 STORY_RILEY2_49,Good luck. Let me know if Felix brings any more surprises in!,ada,,,
-STORY_RILEY2_50,"Definitely. Who would have thought a virtual cat would bring in a virtual mouse? See ya, *{{user_name}}*!",riley,,,
+STORY_RILEY2_50,"Definitely. Who would have thought a virtual cat would bring in a virtual mouse? See ya, <b>{{user_name}}</b>!",riley,,,
