+++ data/quests_strings/story_saniel1.csv
@@ -3,23 +3,23 @@ STORY_SANIEL1_QUEST_SUBTITLE,In this story:,,,,
 STORY_SANIEL1_QUEST_DESCRIPTION,"Visit the Academy Archives and meet Dr. Saniel Rowe, Professor and Head Archivist.",,,,
 STORY_SANIEL1_POSITIVE,👍,,,,
 STORY_SANIEL1_NEGATIVE,👎,,,,
-STORY_SANIEL1_1,"Hello, *{{user_name}}* - I was wondering when you'd stop by.",saniel,,,
+STORY_SANIEL1_1,"Hello, <b>{{user_name}}</b> - I was wondering when you'd stop by.",saniel,,,
 STORY_SANIEL1_2,"I've been trying to get the Academy to search for students with different ways of learning and problem solving - and finally, here you are.",saniel,,,
 STORY_SANIEL1_3,"I want the Academy to be more than just a space for teaching and learning, I want it to be a community, based on sharing knowledge.",saniel,,,
 STORY_SANIEL1_4,"In my past, I have found my deepest friendships and strongest allies to be the result of communities. In all my years of —",saniel,,,
-STORY_SANIEL1_5,"*<CRASH>*
-*<THUD>*",felixnet,,,
+STORY_SANIEL1_5,"<b>&lt;CRASH&gt;</b>
+<b>&lt;THUD&gt;</b>",felixnet,,,
 STORY_SANIEL1_6,What was that?,saniel,,,
 STORY_SANIEL1_7,"Ah, I see. Another pile of books sacrificed to bored cats.",saniel,,,
 STORY_SANIEL1_8,"I should really stop stacking books where the cats like to nap and play. If you see any cats around the Academy, this is where they live - I seem to accumulate cats the way others accumulate art, or souveniers.",saniel,,,
 STORY_SANIEL1_9,"Now, as I was saying - Communities. Communities preserve knowledge and expertise, and our Academy Archives are a good example.",saniel,,,
 STORY_SANIEL1_10,"The Archives are always open to any student - with the exception of the Restricted Stacks. Some things are too easily misused. That said, a simple request will usually open them up... with my supervision.",saniel,,,
-STORY_SANIEL1_11,"Yeah, but supervision means you can't look at anything *really exciting*!",riley,,,
-STORY_SANIEL1_12,"If you consider suppressed computational theorems to be *really exciting*, Riley, then I have severely misjudged you.",saniel,,,
+STORY_SANIEL1_11,"Yeah, but supervision means you can't look at anything <b>really exciting</b>!",riley,,,
+STORY_SANIEL1_12,"If you consider suppressed computational theorems to be <b>really exciting</b>, Riley, then I have severely misjudged you.",saniel,,,
 STORY_SANIEL1_13,"...both for your skill in mathematics and for your preferred reading material, not to mention your knack for… physical intrusion.",saniel,,,
 STORY_SANIEL1_14,Um… Does that mean I'm in trouble?,riley,,,
 STORY_SANIEL1_15,"I would be quite a poor teacher if I punished you for 'breaking in' so easily. If you can open a door by pressing the handle in, then lightly kicking the bottom, it's not very secure.",saniel,,,
-STORY_SANIEL1_16,"… Do you have cameras everywhere? Wait, then you *also* know that there's nothing in that room but drawers full of square plastic cards with holes in the middle!",riley,,,
+STORY_SANIEL1_16,"… Do you have cameras everywhere? Wait, then you <b>also</b> know that there's nothing in that room but drawers full of square plastic cards with holes in the middle!",riley,,,
 STORY_SANIEL1_17,"I must confess I was a step ahead of you, Riley. I moved the Restricted Stacks a month or so ago, but forgot to change the sign.",saniel,,,
 STORY_SANIEL1_18,"Those 'plastic cards' are called floppy disks, by the way. There are several types, and they're used for data storage. Old, but useful.",saniel,,,
 STORY_SANIEL1_19,"Oh, cool. I wish I knew what was on them.",riley,,,
@@ -34,5 +34,5 @@ STORY_SANIEL1_26,"Woah. That's… that's way beyond next-level. If anyone else h
 STORY_SANIEL1_27,"Trust, but verify, Riley - so I'll be expecting a short paper on the method of loci... next week?",saniel,,,
 STORY_SANIEL1_28,"Well, I asked for it. Sounds fun, though!",riley,,,
 STORY_SANIEL1_29,"""Research is the backbone of inquiry,"" as one of my professors once said. Now, I have to get back to work.",saniel,,,
-STORY_SANIEL1_30,"Oh, and Riley - They're called the *Restricted Stacks* for a reason. Some doors are meant to stay locked.",saniel,,,
+STORY_SANIEL1_30,"Oh, and Riley - They're called the <b>Restricted Stacks</b> for a reason. Some doors are meant to stay locked.",saniel,,,
 STORY_SANIEL1_31,"I would tell you both to stay out of trouble, but that would be like asking water to flow uphill. Instead, I'll just say 'see you soon"" and leave it at that.",saniel,,,
